+++ maintenance-api/src/main/java/eu/kennytv/maintenance/api/sponge/MaintenanceSpongeAPI.java
@@ -42,9 +42,9 @@ public final class MaintenanceSpongeAPI {
      * @throws IllegalArgumentException if using a custom (or broken?) version of the plugin, that can't be identified
      */
     public static IMaintenance getAPI() {
-        final Optional<PluginContainer> container = Sponge.getPluginManager().getPlugin("maintenancesponge");
+        final Optional<PluginContainer> container = Sponge.getPluginManager().getPlugin("maintenance");
         if (!container.isPresent() || !container.get().getInstance().isPresent())
-            throw new IllegalArgumentException("Could not get instance of MaintenanceSponge! Broken/custom build of the plugin?");
+            throw new IllegalArgumentException("Could not get instance of Maintenance! Broken/custom build of the plugin?");
         return (IMaintenance) container.get().getInstance().get();
     }
 }